Removed s comments by using conftest file
