+++ sdk/python/tests/utils/data_source_utils.py
@@ -1,53 +1,10 @@
 import contextlib
 import tempfile
-from datetime import datetime, timedelta
 
-import pandas as pd
-import pytest
 from google.cloud import bigquery
 
 from feast.data_format import ParquetFormat
 from feast.data_source import BigQuerySource, FileSource
-
-
-@pytest.fixture
-def simple_dataset_1() -> pd.DataFrame:
-    now = datetime.utcnow()
-    ts = pd.Timestamp(now).round("ms")
-    data = {
-        "id": [1, 2, 1, 3, 3],
-        "float_col": [0.1, 0.2, 0.3, 4, 5],
-        "int64_col": [1, 2, 3, 4, 5],
-        "string_col": ["a", "b", "c", "d", "e"],
-        "ts_1": [
-            ts,
-            ts - timedelta(hours=4),
-            ts - timedelta(hours=3),
-            ts - timedelta(hours=2),
-            ts - timedelta(hours=1),
-        ],
-    }
-    return pd.DataFrame.from_dict(data)
-
-
-@pytest.fixture
-def simple_dataset_2() -> pd.DataFrame:
-    now = datetime.utcnow()
-    ts = pd.Timestamp(now).round("ms")
-    data = {
-        "id": ["a", "b", "c", "d", "e"],
-        "float_col": [0.1, 0.2, 0.3, 4, 5],
-        "int64_col": [1, 2, 3, 4, 5],
-        "string_col": ["a", "b", "c", "d", "e"],
-        "ts_1": [
-            ts,
-            ts - timedelta(hours=4),
-            ts - timedelta(hours=3),
-            ts - timedelta(hours=2),
-            ts - timedelta(hours=1),
-        ],
-    }
-    return pd.DataFrame.from_dict(data)
 
 
 @contextlib.contextmanager
